services/wallet/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
pika>=1.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
requests>=2.31.0 